# Use protocol provided buffer for requests (fixes #1157)