@AudriusButkevicius 2015-07-29
@st-jenkins run integration tests please