@calmh 2015-07-30
What's the advantage of the `bytepool` package over a plain `sync.Pool`?