@AudriusButkevicius 2015-07-30
Not a lot, it deals with buffers of a specific size.