@AudriusButkevicius 2015-07-30
@st-jenkins run integration tests please