@Zillode 2015-07-30
Before: (compares http://build.syncthing.net/view/Testing/job/syncthing-pr/804/console to http://build.syncthing.net/job/syncthing-pr/805/console)
```
Benchmark10kReplace                20     60388016 ns/op
Benchmark10kUpdateChg               5    306595050 ns/op
Benchmark10kUpdateSme              20     88851785 ns/op
Benchmark10kNeed2k                 30     38797488 ns/op
Benchmark10kHaveFullList          100     14241885 ns/op
Benchmark10kGlobal                 10    100060432 ns/op
ok  	github.com/syncthing/syncthing/internal/db	12.418s

BenchmarkMatch              50000        36224 ns/op
BenchmarkMatchCached     10000000          189 ns/op
ok  	github.com/syncthing/syncthing/internal/ignore	4.293s

BenchmarkIndex_10000                      100     17342435 ns/op      2734011 B/op        70135 allocs/op
BenchmarkIndex_100                       2000       813894 ns/op        35867 B/op          821 allocs/op
BenchmarkIndexUpdate_10000_10000           50     31121292 ns/op      4643115 B/op        90139 allocs/op
BenchmarkIndexUpdate_10000_100           1000      1401426 ns/op        52531 B/op         1009 allocs/op
BenchmarkIndexUpdate_10000_1             1000      1296265 ns/op         8068 B/op          117 allocs/op
BenchmarkRequest                     20000000         94.8 ns/op
BenchmarkTree_10000_50                      5    248587136 ns/op    122907361 B/op       624169 allocs/op
BenchmarkTree_100_50                      200      8638800 ns/op      1094204 B/op         5859 allocs/op
BenchmarkTree_100_10                      200      7568059 ns/op       253728 B/op         1853 allocs/op
BenchmarkJobQueueBump                   30000       149529 ns/op
BenchmarkJobQueuePushPopDone10k           500      3507075 ns/op
ok  	github.com/syncthing/syncthing/internal/model	27.279s
```
Pr:
```
Benchmark10kReplace                20     61793258 ns/op
Benchmark10kUpdateChg               5    309457152 ns/op
Benchmark10kUpdateSme              20     89202933 ns/op
Benchmark10kNeed2k                 30     39353746 ns/op
Benchmark10kHaveFullList          100     14162783 ns/op
Benchmark10kGlobal                 10    100099067 ns/op
ok  	github.com/syncthing/syncthing/internal/db	12.498s

BenchmarkMatch              50000        36055 ns/op
BenchmarkMatchCached     10000000          194 ns/op
ok  	github.com/syncthing/syncthing/internal/ignore	4.337s

BenchmarkIndex_10000                      100     17591036 ns/op      2734011 B/op        70135 allocs/op
BenchmarkIndex_100                       2000       831170 ns/op        35630 B/op          821 allocs/op
BenchmarkIndexUpdate_10000_10000           50     32324693 ns/op      4643097 B/op        90139 allocs/op
BenchmarkIndexUpdate_10000_100           1000      1396746 ns/op        52530 B/op         1009 allocs/op
BenchmarkIndexUpdate_10000_1             1000      1288470 ns/op         8068 B/op          117 allocs/op
BenchmarkRequest                     20000000         94.7 ns/op
BenchmarkTree_10000_50                      5    246361569 ns/op    122910184 B/op       624179 allocs/op
BenchmarkTree_100_50                      200      8912803 ns/op      1094195 B/op         5859 allocs/op
BenchmarkTree_100_10                      200      8011735 ns/op       253740 B/op         1853 allocs/op
BenchmarkJobQueueBump                   30000       144459 ns/op
BenchmarkJobQueuePushPopDone10k           500      3471778 ns/op
ok  	github.com/syncthing/syncthing/internal/model	27.367s
```