@AudriusButkevicius 2015-07-30
These don't test the part I am changing sensibly. A good benchmark would be a graph of memory usage while transferring a 5GB file which perhaps can be done via: https://github.com/syncthing/syncthing/blob/master/test/transfer-bench_test.go