@calmh 2015-08-02
@st-jenkins run integration tests please