@AudriusButkevicius 2015-08-02
Did you run any benchmakrs by any chanse?